fixing flow errors. moving get-client-rect to src
