+++ test/unit/state/hook-middleware.spec.js
@@ -2,7 +2,7 @@
 import middleware from '../../../src/state/hook-middleware';
 import { getDraggableDimension, getDroppableDimension } from '../../../src/state/dimension';
 import { clean } from '../../../src/state/action-creators';
-import getClientRect from '../../utils/get-client-rect';
+import getClientRect from '../../../src/state/get-client-rect';
 import noImpact from '../../../src/state/no-impact';
 import type {
   DraggableId,
